[Updated] Air lock entity. Requirements: 2 models: 1) the model that's supposed to be open in the end (with the animations) 2) the model that's supposed to be closed (modelname for the normal model + E(probably other letter in the future)
In the list the E model needs to be registered!

Updated working (due to problems with the collision models).

The Entity starts as the closed model, once the player presses use on it the model changes to the one with animations and performs them.

Once the player then presses use on it again the apropriate animation will be run and then the model will be changed to the E one again.